Reduce VC++ warning load from 5000-odd to 529
git-svn-id: file:///openmsx-svn/code/openmsx/trunk@8536 7512d703-071c-0410-b5c2-d2bdd895b7d1

+++ src/serial/MidiInWindows.cc
@@ -142,7 +142,7 @@ void MidiInWindows::run()
 				break;
 			case MM_MIM_DATA:
 			case MM_MIM_MOREDATA:
-				procShortMsg(msg.lParam);
+				procShortMsg(DWORD(msg.lParam));
 				break;
 			case MM_MIM_LONGDATA:
 				procLongMsg(reinterpret_cast<LPMIDIHDR>(msg.lParam));
